More unused files removed.
