--- src/services/koulutukset/actionTypes.js
@@ -1,11 +1,0 @@
-export const FETCH_KOULUTUKSET_ALL_START = "FETCH_KOULUTUKSET_ALL_START";
-export const FETCH_KOULUTUKSET_ALL_SUCCESS = "FETCH_KOULUTUKSET_ALL_SUCCESS";
-export const FETCH_KOULUTUKSET_ALL_FAILURE = "FETCH_KOULUTUKSET_ALL_FAILURE";
-
-export const FETCH_KOULUTUKSET_MUUT_START = "FETCH_KOULUTUKSET_MUUT_START";
-export const FETCH_KOULUTUKSET_MUUT_SUCCESS = "FETCH_KOULUTUKSET_MUUT_SUCCESS";
-export const FETCH_KOULUTUKSET_MUUT_FAILURE = "FETCH_KOULUTUKSET_MUUT_FAILURE";
-
-export const FETCH_KOULUTUS_START = "FETCH_KOULUTUS_START";
-export const FETCH_KOULUTUS_SUCCESS = "FETCH_KOULUTUS_SUCCESS";
-export const FETCH_KOULUTUS_FAILURE = "FETCH_KOULUTUS_FAILURE";
